requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.5.2
scipy>=1.9.0
statsmodels>=0.13.0
category_encoders>=2.6.3
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.13.0
bokeh>=3.3.0
altair>=5.2.0
xgboost>=2.0.0
shap>=0.44.0
pyarrow>=14.0.1
openpyxl>=3.1.2
openai>=1.0.0
anthropic>=0.3.0
cryptography>=40.0.0
psutil>=5.9.0
requests>=2.28.0
wordcloud>=1.9.0
yfinance>=0.2.35
geopy>=2.3.0
folium>=0.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tabulate
PySide6[webengine]
python-louvain>=0.16
networkx>=3.2 
typing_extensions>=4.9.0
pyinstaller>=5.13.0
kaleido
httpx>=0.26.0